- Move Api services to services/api directory. - Remove initial calls for setActive method in peers service
